services/router-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
PyYAML==6.0.2
redis==5.0.8
prometheus-client==0.20.0
python-json-logger==2.0.7
